Implemented routing and restructured project
 - Added in React Router/react-router-redux and set up a basic routing architecture
   - Created a skeleton not found page that matches all non-defined routes
 - Moved `App.js` to `components/product_cards/ProductCards.js` to make way for additional pages in the future
 - Created a `helpers` directory and moved some files into it that are shared by multiple parts of the application
 - Moved all components related to the product page into the `product_page` subdirectory
