+++ src/components/NotFound.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import  { Container, Heading } from '../helpers/container';
+
+const NotFound = () => {
+  return (
+    <Container>
+      <Heading>Member Application</Heading>
+    </Container>
+  );
+}
+
+export default NotFound;
