+++ src/index.js
@@ -7,19 +7,19 @@ import { createStore } from 'redux';
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 import injectTapEventPlugin from 'react-tap-event-plugin';
 
-import App from './App';
 import reducers from './reducers';
 
 import registerServiceWorker from './registerServiceWorker';
 import './index.css';
+import CoreRouter from './router';
 
 injectTapEventPlugin(); //needed for material-ui
 
 const Root = ()=>{
-  return(
+  return (
     <Provider store={createStore(reducers)}>
       <MuiThemeProvider>
-        <App />
+        <CoreRouter />
       </MuiThemeProvider>
     </Provider>
   );
